WIP: db integration tests

+++ storage/category/category.go
@@ -4,7 +4,7 @@ const (
 	// common categories
 
 	// Default is the ID of category, that must be used for all new imported transactions.
-	Default = 1
+	Default = int32(1)
 	// Ignored is the ID of category, that would be ignored in report.
-	Ignored = 127
+	Ignored = int32(127)
 )
